Russ added revised PieceMove, Turn, and RollDice
Russ added revised PieceMove, Turn, and RollDice classes to agree with Rev 2 UML class diagram.

Note these revisions suggest changes to be made to Ramiro's PieceMove, Turn, and RollDice classes to interface with the 
Tile, Piece, Player, and UrGameBoard classes that I have set up. 

More editing is needed to complete, but I want to capture the code we have done so far.